divide method completed better

+++ src/main/java/ie/atu/week3/Calc.java
@@ -7,19 +7,19 @@ public class Calc {
         //methods
         System.out.println("Please enter number 1: ");
         Scanner sc = new Scanner(System.in);
-        int num1 = sc.nextInt();
+        float num1 = sc.nextFloat();
         System.out.println("Please enter number 2: ");
-        int num2 = sc.nextInt();
+        float num2 = sc.nextFloat();
         System.out.println("Please enter operator: ");
         String op = sc.next();
 
 
         switch(op) {
-            case "add": add(num1, num2);break;
+            case "add": add((int) num1, (int)num2);break;
 
-            case "subtract": subtract(num1, num2);break;
+            case "subtract": subtract((int) num1, (int)num2);break;
 
-            case "multiply": multiply(num1, num2);break;
+            case "multiply": multiply((int) num1, (int)num2);break;
 
             case "divide": divide(num1, num2);break;
 
@@ -42,14 +42,14 @@ public class Calc {
         System.out.println("The sum is: " + sum);
     }
 
-    public static void divide(int a,int b) {
+    public static void divide(float a,float b) {
 
         if(b==0){
             System.out.println("Division by zero");
             return;
         }
         else {
-            int sum = a/b;
+            float sum = a/b;
             System.out.println("The sum is: " + sum);
         }
     }
